Update project name and version, improve AABB structure, and enhance logging and CRC classes

+++ include/splat/maths/maths.h
@@ -48,7 +48,7 @@ namespace splat {
  */
 template <typename T>
 inline T sigmoid(T x) {
-  return 1.0 / (1.0 + exp(-x));
+  return 1.0 / (1.0 + std::exp(-static_cast<double>(x)));
 }
 
 /**
